--- models/formModels/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    Form: require('./Form')
-};
-
