+++ models/index.js
@@ -1,6 +1,6 @@
 const User = require("./User");
 const Resume = require("./Resume");
-const Form = require("./formModels/Form");
+const Form = require("./Form");
 const Signup = require("./Signup");
 
 module.exports = { User, Resume, Form, Signup };
